feat(mobile): wired up member settings button
