+++ apps/mobile/app/[site_id]/(tabs)/home/quick-search.tsx
@@ -62,7 +62,6 @@ export default function QuickSearch() {
                 <SearchInput
                     style={{ backgroundColor: colors.grey5 }}
                     placeholder="Search"
-                    iconColor={colors.destructive}
                     placeholderTextColor={colors.grey}
                     onChangeText={setSearchQuery}
                     value={searchQuery}
