+++ apps/mobile/app/[site_id]/chat/[id]/channel-settings.tsx
@@ -60,7 +60,7 @@ const ChannelSettings = () => {
                 <View className="flex-col gap-5">
                     <ChannelBaseDetails channelData={channelData} />
                     <Divider className='mx-0' prominent />
-                    <MembersTray onViewAll={() => router.push(`../channel-settings-members`, { relativeToDirectory: true })} />
+                    <MembersTray onViewAll={() => router.push(`../channel-members`, { relativeToDirectory: true })} />
                     <Divider className='mx-0' prominent />
                     <View className="flex-col gap-2">
                         <Text className="text-[15px] font-medium px-4">Settings</Text>
